pretty sure everything is fine
writes into index the *delete*, and in get contents scans for delete and edit calls, if finds then doesn't add them to return array and instead to the delete or edit array. After adding everything else, deletes it if it is in the current staged things. Else, do >>> traversal things starting with the tree in the array